+++ blinksy/src/driver/clockless/mod.rs
@@ -208,6 +208,7 @@ pub trait ClocklessWriterAsync<Led: ClocklessLed> {
 /// - `Led` - The LED protocol implementation (must implement ClocklessLed)
 /// - `Writer` - The clocked writer
 #[derive(Debug)]
+#[cfg_attr(feature = "defmt", derive(defmt::Format))]
 pub struct ClocklessDriver<Led, Writer> {
     /// Marker for the LED protocol type
     led: PhantomData<Led>,